Move `StateScheduler` to a new shared sub-package
