+++ denotational/core/src/main/kotlin/io/github/oxidefrp/core/Cell.kt
@@ -1,5 +1,7 @@
 package io.github.oxidefrp.core
 
+import io.github.oxidefrp.core.shared.StateSchedulerLayer
+
 data class ValueChange<out A>(
     val oldValue: A,
     val newValue: A,
