+++ denotational/core/src/main/kotlin/io/github/oxidefrp/core/MomentState.kt
@@ -1,5 +1,7 @@
 package io.github.oxidefrp.core
 
+import io.github.oxidefrp.core.shared.StateSchedulerLayer
+
 abstract class MomentState<S, out A> {
     companion object {
         fun <S, A> pure(value: A): MomentState<S, A> =
