+++ denotational/core/src/test/kotlin/io/github/oxidefrp/core/StateStructureDenotationalUnitTests.kt
@@ -1,5 +1,6 @@
 package io.github.oxidefrp.core
 
+import io.github.oxidefrp.core.shared.StateSchedulerLayer
 import io.github.oxidefrp.core.test_utils.tableSignal
 import kotlin.test.Test
 import kotlin.test.assertEquals
